+++ src/main/java/com/service/IService.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface IService<T> {
     List<T> findAll();
 
-    T findById(int id);
+    T findById(Long id);
 
     void save(T t);
 
-    void remove(int id);
+    void remove(Long id);
 }
